steamAwardNominateGame_v2
add more save way to get the emblem. this will use the search. to find a game for each category, to ensure it is valid. and will ensure to get a game that have over 5 min of game play.
this will be better, if you have a lot of account that doent have the same games.