add active free game, clean wishlist and steam IPs
When a game is 100% free like app 325180 ( as this commid was made it was 100%  free ) will be active on the account.
Added do it will clean the wishlist.
Add steam Ips, as the current steam node ip, do not exist, so sometimes it can fail on login. now it shoud not.